Moved V8BridgeInvokeTests to shared-tests module
