+++ platforms/android/bridge-v8/src/androidTest/AndroidManifest.xml
@@ -1,12 +1,3 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.salesforce.nimbus.bridge.v8">
-
-    <uses-permission android:name="android.permission.INTERNET" />
-
-    <application
-        android:label="@string/app_name">
-        <activity android:name=".WebViewActivity"/>
-    </application>
-
-</manifest>
+<manifest
+    package="com.salesforce.nimbus.bridge.v8"/>
