--- platforms/android/bridge-v8/src/androidTest/java/com/salesforce/nimbus/bridge/v8/WebViewActivity.kt
@@ -1,18 +1,0 @@
-//
-// Copyright (c) 2019, Salesforce.com, inc.
-// All rights reserved.
-// SPDX-License-Identifier: BSD-3-Clause
-// For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
-//
-
-package com.salesforce.nimbus.bridge.v8
-
-import android.app.Activity
-import android.os.Bundle
-
-class WebViewActivity : Activity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-    }
-}
