+++ platforms/android/shared-tests/src/androidTest/java/com/salesforce/nimbus/bridge/tests/V8BridgeInvokeTests.kt
@@ -1,8 +1,10 @@
-package com.salesforce.nimbus.bridge.v8
+package com.salesforce.nimbus.bridge.tests
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.eclipsesource.v8.V8
 import com.google.common.truth.Truth.assertThat
+import com.salesforce.nimbus.bridge.v8.V8Bridge
+import com.salesforce.nimbus.bridge.v8.toV8Encodable
 import com.salesforce.nimbus.invoke
 import com.salesforce.nimbus.k2v8.scope
 import kotlinx.serialization.Serializable
@@ -10,8 +12,6 @@ import org.junit.After
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
-import java.util.concurrent.CountDownLatch
-import java.util.concurrent.TimeUnit
 
 /**
  * Tests the [V8Bridge.invoke] function.
@@ -79,7 +79,12 @@ class V8BridgeInvokeTests {
                 bridge.invoke("objectFunc", emptyArray(), SomeClass.serializer()) { error, result ->
                     assertThat(error).isNull()
                     assertThat(result).isNotNull()
-                    assertThat(result).isEqualTo(SomeClass("Some string", 5))
+                    assertThat(result).isEqualTo(
+                        SomeClass(
+                            "Some string",
+                            5
+                        )
+                    )
                     countDown()
                 }
             }
@@ -109,12 +114,6 @@ class V8BridgeInvokeTests {
                     countDown()
                 }
             }
-        }
-    }
-
-    private fun withinLatch(block: CountDownLatch.() -> Unit): CountDownLatch {
-        return CountDownLatch(1).apply { block(this) }.also {
-            assertThat(it.await(5, TimeUnit.SECONDS)).isTrue()
         }
     }
 }
